+++ core/edge-segments.h
@@ -34,7 +34,7 @@ public:
 	LinearSegment(Point2 p0, Point2 p1);
 	Point2 point(double param) const;
 	Vector2 direction(double param) const;
-	float signedDistance(Point2 origin) const;
+	double signedDistance(Point2 origin) const;
 	void bounds(double &l, double &b, double &r, double &t) const;
 
 	void moveStartPoint(Point2 to);
@@ -56,7 +56,7 @@ public:
 	QuadraticSegment(Point2 p0, Point2 p1, Point2 p2);
 	Point2 point(double param) const;
 	Vector2 direction(double param) const;
-	float signedDistance(Point2 origin) const;
+    double signedDistance(Point2 origin) const;
 	void bounds(double &l, double &b, double &r, double &t) const;
 
 	void moveStartPoint(Point2 to);
@@ -78,7 +78,7 @@ public:
 	CubicSegment(Point2 p0, Point2 p1, Point2 p2, Point2 p3);
 	Point2 point(double param) const;
 	Vector2 direction(double param) const;
-	float signedDistance(Point2 origin) const;
+    double signedDistance(Point2 origin) const;
 	void bounds(double &l, double &b, double &r, double &t) const;
 
 	void moveStartPoint(Point2 to);
@@ -136,7 +136,7 @@ public:
 		RETBYTYPE(direction, param);
 	}
 	/// Returns the minimum signed distance between origin and the edge.
-	float signedDistance(Point2 origin) const
+    double signedDistance(Point2 origin) const
 	{
 		RETBYTYPE(signedDistance, origin);
 	}
